Project restrucurize. Nbp API with webclient init.
