+++ Backend/pom.xml
@@ -111,6 +111,11 @@
 			<version>0.12.6</version>
 			<scope>runtime</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter</artifactId>
+			<version>3.5.3</version>
+		</dependency>
 	</dependencies>
 	<build>
 		<plugins>
