+++ Backend/src/main/java/com/financialtransactions/monitor/domain/enums/CurrencyType.java
@@ -1,0 +1,7 @@
+package com.financialtransactions.monitor.domain.enums;
+
+public enum CurrencyType {
+    PLN,
+    USD,
+    EUR
+}
