+++ Backend/src/main/java/com/financialtransactions/monitor/domain/enums/TradeType.java
@@ -1,0 +1,5 @@
+package com.financialtransactions.monitor.domain.enums;
+
+public enum TradeType {
+    BUY, SELL
+}
