--- Backend/src/main/java/com/financialtransactions/monitor/model/CurrencyType.java
@@ -1,7 +1,0 @@
-package com.financialtransactions.monitor.model;
-
-public enum CurrencyType {
-    PLN,
-    USD,
-    EUR
-}
