--- Backend/src/main/java/com/financialtransactions/monitor/model/TradeType.java
@@ -1,5 +1,0 @@
-package com.financialtransactions.monitor.model;
-
-public enum TradeType {
-    BUY, SELL
-}
