+++ Backend/src/main/java/com/financialtransactions/monitor/modules/fund/dto/FundDto.java
@@ -1,0 +1,28 @@
+package com.financialtransactions.monitor.modules.fund.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.financialtransactions.monitor.domain.enums.CurrencyType;
+import lombok.*;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+@Builder
+public class FundDto {
+
+    private Long id;
+    private String symbol;
+    private String name;
+    private CurrencyType currencyType;
+    private String market;
+    private String sector;
+    private String exchange;
+    private BigDecimal currentPrice;
+
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    private LocalDateTime lastUpdated;
+}
