+++ Backend/src/main/java/com/financialtransactions/monitor/modules/portfolio/service/PortfolioService.java
@@ -1,0 +1,10 @@
+package com.financialtransactions.monitor.modules.portfolio.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class PortfolioService {
+
+}
