--- Backend/src/main/java/com/financialtransactions/monitor/repository/FundRepository.java
@@ -1,10 +1,0 @@
-package com.financialtransactions.monitor.repository;
-
-import com.financialtransactions.monitor.model.Fund;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface FundRepository extends JpaRepository<Fund, Long> {
-
-}
