--- Backend/src/main/java/com/financialtransactions/monitor/service/PortfolioService.java
@@ -1,10 +1,0 @@
-package com.financialtransactions.monitor.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class PortfolioService {
-
-}
